refactor(routes): :art: Added User route handler
Nested getUser under this, only grouping getUsers on folder level

Using `/:id` on user router could match unintended routes hence it should be placed last.
